Conflicts:
	modules/vlc_audio_video
	tao/xlr

+++ fresnel.cpp
@@ -116,7 +116,7 @@ void Fresnel::Draw()
         // Get and set camera position
         Vector3 cam;
         tao->getCamera(&cam, NULL, NULL, NULL);
-        GLfloat camera[3] = {cam.x, cam.y, cam.z};
+        GLfloat camera[3] = { (float) cam.x, (float) cam.y, (float) cam.z};
         glUniform3fv(uniforms["camera"], 1, camera);
 
         if(tao->isGLExtensionAvailable("GL_EXT_gpu_shader4"))
